refactor: extract retrieval of podcast index auth manager
