+++ backend/service/podcastEpisodeService.ts
@@ -1,23 +1,5 @@
-import { PodcastIndexAuthManager } from "../api/authManager.js"
+import { getPodcastIndexAuthManager } from "../api/authManager.js"
 import { PodcastIndexFacade } from "../api/podcastFacade.js"
-import { InvalidApiKeyError } from "../error/invalidApiKeyError.js"
-
-function getPodcastIndexAuthManager() {
-  const apiKey = process.env.PODCAST_INDEX_API_KEY
-  const apiSecret = process.env.PODCAST_INDEX_API_SECRET
-  if (
-    apiKey == null ||
-    apiKey === "" ||
-    apiSecret == null ||
-    apiSecret === ""
-  ) {
-    throw new InvalidApiKeyError(
-      "Server configuration error: Invalid Podcast API Key"
-    )
-  }
-  const podcastAuthManager = new PodcastIndexAuthManager(apiKey, apiSecret)
-  return podcastAuthManager
-}
 
 export async function getPodcastEpisodes(podcastId: string, limit: number) {
   const podcastAuthManager = getPodcastIndexAuthManager()
